requirements.txt
gitdb2==2.0.5
GitPython==2.1.11
pygerrit2==2.0.9
